+++ src/main/java/insane96mcp/buildersconstruct/modifiers/AngelBuilderModifier.java
@@ -9,12 +9,16 @@ import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.client.renderer.LevelRenderer;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.core.BlockPos;
+import net.minecraft.sounds.SoundSource;
 import net.minecraft.util.Mth;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.ClipContext;
+import net.minecraft.world.level.block.SoundType;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.HitResult;
 import net.minecraft.world.phys.Vec2;
 import net.minecraft.world.phys.Vec3;
@@ -22,6 +26,7 @@ import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.client.event.RenderLevelStageEvent;
+import net.minecraftforge.common.ForgeMod;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import slimeknights.tconstruct.library.modifiers.ModifierEntry;
@@ -38,6 +43,8 @@ import slimeknights.tconstruct.library.tools.nbt.ToolStack;
 
 @Mod.EventBusSubscriber(modid = BuildersConstruct.MOD_ID)
 public class AngelBuilderModifier extends NoLevelsModifier implements GeneralInteractionModifierHook {
+
+    private static final double BASE_DISTANCE = 2.5d;
 
     @Override
     protected void registerHooks(ModifierHookMap.Builder hookBuilder) {
@@ -61,30 +68,39 @@ public class AngelBuilderModifier extends NoLevelsModifier implements GeneralInt
         if (!player.getLevel().isClientSide) {
             Vec2 vec2 = player.getRotationVector();
             Vec3 vec3 = player.getEyePosition();
-            float yCos = Mth.cos((vec2.y + 90.0F) * ((float)Math.PI / 180F));
-            float ySin = Mth.sin((vec2.y + 90.0F) * ((float)Math.PI / 180F));
-            float xCos = Mth.cos(-vec2.x * ((float)Math.PI / 180F));
-            float xSin = Mth.sin(-vec2.x * ((float)Math.PI / 180F));
-            Vec3 forwardVec = new Vec3(yCos * xCos, xSin, ySin * xCos);
-            double d0 = forwardVec.x * 3;
-            double d1 = forwardVec.y * 3;
-            double d2 = forwardVec.z * 3;
-            Vec3 endRayCast = new Vec3(vec3.x + d0, vec3.y + d1, vec3.z + d2);
+            double reach = player.getAttributeValue(ForgeMod.REACH_DISTANCE.get()) - player.getAttributeBaseValue(ForgeMod.REACH_DISTANCE.get());
+            Vec3 endRayCast = getEndRayCast(vec2, vec3, reach);
             HitResult hitResult = player.level.clip(new ClipContext(player.getEyePosition(), endRayCast, ClipContext.Block.OUTLINE, ClipContext.Fluid.ANY, null));
             if (hitResult.getType() == HitResult.Type.BLOCK)
                 return InteractionResult.PASS;
             BlockPos pos = new BlockPos(endRayCast);
             if (!player.level.getBlockState(pos).getMaterial().isReplaceable())
                 return InteractionResult.PASS;
-            player.level.setBlock(pos, blockItemToPlace.getBlock().defaultBlockState(), 3);
-            if (!player.getAbilities().instabuild) {
-                player.getOffhandItem().shrink(1);
-            }
-            if (ToolDamageUtil.directDamage(tool, 20, player, player.getItemInHand(hand))) {
-                player.broadcastBreakEvent(hand);
+            BlockPlaceContext blockPlaceContext = new BlockPlaceContext(player, InteractionHand.OFF_HAND, player.getItemInHand(InteractionHand.OFF_HAND), ConstructionModifier.createBlockHitResult(endRayCast, player.getDirection(), pos));
+            InteractionResult interactionResult = blockItemToPlace.place(blockPlaceContext);
+            if (interactionResult.consumesAction()) {
+                BlockState placedState = player.level.getBlockState(blockPlaceContext.getClickedPos());
+                SoundType soundtype = placedState.getSoundType(player.level, blockPlaceContext.getClickedPos(), player);
+                player.level.playSound(null, blockPlaceContext.getClickedPos(), placedState.getSoundType(player.level, blockPlaceContext.getClickedPos(), player).getPlaceSound(), SoundSource.BLOCKS, (soundtype.getVolume() + 1.0F) / 2.0F, soundtype.getPitch() * 0.8F);
+
+                if (ToolDamageUtil.directDamage(tool, 20, player, player.getItemInHand(hand))) {
+                    player.broadcastBreakEvent(hand);
+                }
             }
         }
         return InteractionResult.sidedSuccess(player.level.isClientSide);
+    }
+
+    private static Vec3 getEndRayCast(Vec2 rotation, Vec3 startingPos, double bonusReach) {
+        float yCos = Mth.cos((rotation.y + 90.0F) * ((float) Math.PI / 180F));
+        float ySin = Mth.sin((rotation.y + 90.0F) * ((float) Math.PI / 180F));
+        float xCos = Mth.cos(-rotation.x * ((float) Math.PI / 180F));
+        float xSin = Mth.sin(-rotation.x * ((float) Math.PI / 180F));
+        Vec3 forwardVec = new Vec3(yCos * xCos, xSin, ySin * xCos);
+        double d0 = forwardVec.x * (BASE_DISTANCE + bonusReach);
+        double d1 = forwardVec.y * (BASE_DISTANCE + bonusReach);
+        double d2 = forwardVec.z * (BASE_DISTANCE + bonusReach);
+        return new Vec3(startingPos.x + d0, startingPos.y + d1, startingPos.z + d2);
     }
 
     @OnlyIn(Dist.CLIENT)
@@ -109,15 +125,8 @@ public class AngelBuilderModifier extends NoLevelsModifier implements GeneralInt
             return;
         Vec2 vec2 = player.getRotationVector();
         Vec3 vec3 = player.getEyePosition();
-        float yCos = Mth.cos((vec2.y + 90.0F) * ((float)Math.PI / 180F));
-        float ySin = Mth.sin((vec2.y + 90.0F) * ((float)Math.PI / 180F));
-        float xCos = Mth.cos(-vec2.x * ((float)Math.PI / 180F));
-        float xSin = Mth.sin(-vec2.x * ((float)Math.PI / 180F));
-        Vec3 forwardVec = new Vec3(yCos * xCos, xSin, ySin * xCos);
-        double d0 = forwardVec.x * 3;
-        double d1 = forwardVec.y * 3;
-        double d2 = forwardVec.z * 3;
-        Vec3 endRayCast = new Vec3(vec3.x + d0, vec3.y + d1, vec3.z + d2);
+        double reach = player.getAttributeValue(ForgeMod.REACH_DISTANCE.get()) - player.getAttributeBaseValue(ForgeMod.REACH_DISTANCE.get());
+        Vec3 endRayCast = getEndRayCast(vec2, vec3, reach);
         HitResult hitResult = player.level.clip(new ClipContext(player.getEyePosition(), endRayCast, ClipContext.Block.OUTLINE, ClipContext.Fluid.ANY, null));
         if (hitResult.getType() == HitResult.Type.BLOCK)
             return;